`connection.invokeGraphqlAPI` => `graphql.execute`
